+++ Editor/Authoring/CloudContent/CcdCloudStorage.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using System.Net;
 using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Threading;
@@ -16,7 +17,10 @@ using Unity.Services.Multiplay.Authoring.Editor.AdminApis.Ccd.Entries;
 using Unity.Services.Multiplay.Authoring.Editor.AdminApis.Ccd.Http;
 using Unity.Services.Multiplay.Authoring.Editor.AdminApis.Ccd.Models;
 using Unity.Services.Multiplay.Authoring.Core.CloudContent;
+using Unity.Services.Multiplay.Authoring.Editor.MultiplayApis;
+using UnityEngine;
 using HttpClient = System.Net.Http.HttpClient;
+using ILogger = Unity.Services.Multiplay.Authoring.Core.Logging.ILogger;
 
 namespace Unity.Services.Multiplay.Authoring.Editor.CloudContent
 {
@@ -24,44 +28,67 @@ namespace Unity.Services.Multiplay.Authoring.Editor.CloudContent
     {
         const int k_BatchSize = 10;
 
-        readonly ICcdApiConfig m_ApiConfig;
+        IMultiplayApiConfig m_ApiConfig;
         readonly IBucketsApiClient m_BucketsApiClient;
         readonly IEntriesApiClient m_EntriesApiClient;
         readonly HttpClient m_UploadClient;
         readonly ICcdAnalytics m_Analytics;
+        readonly IApiAuthenticator m_ApiInit;
+        private readonly ILogger m_Logger;
 
         public CcdCloudStorage(
-            ICcdApiConfig apiConfig,
             IBucketsApiClient bucketsApiClient,
             IEntriesApiClient entriesApiClient,
             HttpClient uploadClient,
-            ICcdAnalytics analytics)
+            ICcdAnalytics analytics,
+            IApiAuthenticator apiInit,
+            ILogger logger)
         {
-            m_ApiConfig = apiConfig;
             m_BucketsApiClient = bucketsApiClient;
             m_EntriesApiClient = entriesApiClient;
             m_UploadClient = uploadClient;
             m_Analytics = analytics;
+            m_ApiInit = apiInit;
+            m_Logger = logger;
+            m_ApiConfig = ApiConfig.Empty;
+        }
+
+        public async Task InitAsync()
+        {
+            var(config, basePath, headers) = await m_ApiInit.Authenticate();
+            m_ApiConfig = config;
+            ((BucketsApiClient)m_BucketsApiClient).Configuration = new AdminApis.Ccd.Configuration(
+                basePath,
+                null,
+                null,
+                headers);
+            ((EntriesApiClient)m_EntriesApiClient).Configuration = new AdminApis.Ccd.Configuration(
+                basePath,
+                null,
+                null,
+                headers);
         }
 
         public async Task<CloudBucketId> FindBucket(string name)
         {
-            var request = new ListBucketsByProjectEnvRequest(m_ApiConfig.EnvironmentId, m_ApiConfig.ProjectId, name: name);
+            var request = new ListBucketsByProjectEnvRequest(m_ApiConfig.EnvironmentId.ToString(), m_ApiConfig.ProjectId.ToString(), name: name);
             var buckets = await m_BucketsApiClient.ListBucketsByProjectEnvAsync(request);
-            return buckets.Result.Select(b => new CloudBucketId { Id = b.Id }).FirstOrDefault();
+            return buckets.Result.Select(b => new CloudBucketId { Guid = b.Id }).FirstOrDefault();
         }
 
         public async Task<CloudBucketId> CreateBucket(string name)
         {
-            var projectGuid = Guid.Parse(m_ApiConfig.ProjectId);
-            var request = new CreateBucketByProjectEnvRequest(m_ApiConfig.EnvironmentId, m_ApiConfig.ProjectId, new InlineObject32(name, projectGuid));
+            var projectGuid = Guid.Parse(m_ApiConfig.ProjectId.ToString());
+            var request = new CreateBucketByProjectEnvRequest(m_ApiConfig.EnvironmentId.ToString(), m_ApiConfig.ProjectId.ToString(), new InlineObject32(name, projectGuid));
             var res = await m_BucketsApiClient.CreateBucketByProjectEnvAsync(request);
-            return new CloudBucketId { Id = res.Result.Id };
+            return new CloudBucketId { Guid = res.Result.Id };
         }
 
         public async Task<int> UploadBuildEntries(CloudBucketId bucket, IList<BuildEntry> localEntries, Action<BuildEntry> onUpdated = null, CancellationToken cancellationToken = default)
         {
             var changes = 0;
+            var unchanged = 0;
+            var deleted = 0;
             try
             {
                 using var upload = m_Analytics.BeginUpload();
@@ -74,18 +101,29 @@ namespace Unity.Services.Multiplay.Authoring.Editor.CloudContent
                     var(path, content) = local;
                     var normalizedPath = Normalize(path);
                     var hash = content.CcdHash();
-                    if (remoteEntries.ContainsKey(normalizedPath))
+                    var entryExists = remoteEntries.TryGetValue(normalizedPath, out var remoteEntry);
+                    if (entryExists)
                     {
                         orphans.Remove(normalizedPath);
-                        if (remoteEntries[normalizedPath].ContentHash.ToLowerInvariant() == hash)
+                        bool uploadedCorrectly = remoteEntry.Complete;
+                        bool hasChanged = remoteEntry.ContentHash.ToLowerInvariant() != hash;
+                        if (!uploadedCorrectly)
                         {
+                            m_Logger.LogVerbose($"Re-uploading Entry {remoteEntry.Entryid}|{remoteEntry.Path}");
+                        }
+
+                        bool shouldSkip = !hasChanged && uploadedCorrectly;
+                        if (shouldSkip)
+                        {
+                            //Dont reupload entry unchanged entry.
+                            Interlocked.Increment(ref unchanged);
                             onUpdated?.Invoke(local);
                             return;
                         }
                     }
 
                     var entry = await CreateOrUpdateEntry(bucket, normalizedPath, hash, (int)content.Length);
-                    changes++;
+                    Interlocked.Increment(ref changes);
                     await UploadSignedContent(entry, content);
                     onUpdated?.Invoke(local);
                 });
@@ -93,33 +131,36 @@ namespace Unity.Services.Multiplay.Authoring.Editor.CloudContent
                 await orphans.BatchAsync(k_BatchSize, async orphan =>
                 {
                     var entry = remoteEntries[orphan];
-                    changes++;
+                    Interlocked.Increment(ref changes);
+                    Interlocked.Increment(ref deleted);
                     await DeleteEntry(bucket, entry);
                 });
             }
             catch (Exception e)
             {
+                m_Logger.LogVerbose($"Upload failed. Total changes: {changes}; unchanged: {unchanged}; deleted: {deleted}");
                 m_Analytics.UploadFailed(e.GetType().ToString());
                 throw;
             }
 
+            m_Logger.LogVerbose($"Upload succeed. Total changes: {changes}; unchanged: {unchanged}; deleted: {deleted}");
             return changes;
         }
 
         public async Task Clear()
         {
             var buckets = await m_BucketsApiClient.ListBucketsByProjectEnvAsync(new ListBucketsByProjectEnvRequest(
-                m_ApiConfig.EnvironmentId,
-                m_ApiConfig.ProjectId));
+                m_ApiConfig.EnvironmentId.ToString(),
+                m_ApiConfig.ProjectId.ToString()));
 
             foreach (var bucket in buckets.Result)
             {
                 try
                 {
                     await m_BucketsApiClient.DeleteBucketEnvAsync(new DeleteBucketEnvRequest(
-                        m_ApiConfig.EnvironmentId,
+                        m_ApiConfig.EnvironmentId.ToString(),
                         bucket.Id.ToString(),
-                        m_ApiConfig.ProjectId
+                        m_ApiConfig.ProjectId.ToString()
                     ));
                 }
                 catch (HttpException e) when (e.Response.StatusCode == 403)
@@ -133,7 +174,11 @@ namespace Unity.Services.Multiplay.Authoring.Editor.CloudContent
 
         async Task DeleteEntry(CloudBucketId bucket, InlineResponse2003 entry)
         {
-            var deleteReq = new DeleteEntryEnvRequest(m_ApiConfig.EnvironmentId, bucket.ToString(), entry.Entryid.ToString(), m_ApiConfig.ProjectId);
+            var deleteReq = new DeleteEntryEnvRequest(
+                m_ApiConfig.EnvironmentId.ToString(),
+                bucket.ToString(),
+                entry.Entryid.ToString(),
+                m_ApiConfig.ProjectId.ToString());
             await m_EntriesApiClient.DeleteEntryEnvAsync(deleteReq);
         }
 
@@ -141,10 +186,10 @@ namespace Unity.Services.Multiplay.Authoring.Editor.CloudContent
         {
             var create = new InlineObject23(hash, length, signedUrl: true);
             var request = new CreateOrUpdateEntryByPathEnvRequest(
-                m_ApiConfig.EnvironmentId,
+                m_ApiConfig.EnvironmentId.ToString(),
                 bucket.ToString(),
                 path,
-                m_ApiConfig.ProjectId,
+                m_ApiConfig.ProjectId.ToString(),
                 create,
                 updateIfExists: true);
             var res = await m_EntriesApiClient.CreateOrUpdateEntryByPathEnvAsync(request);
@@ -171,8 +216,22 @@ namespace Unity.Services.Multiplay.Authoring.Editor.CloudContent
             var page = 1;
             do
             {
-                var request = new GetEntriesEnvRequest(m_ApiConfig.EnvironmentId, bucket.ToString(), m_ApiConfig.ProjectId, page, perPage: entriesPerPage);
-                res = await m_EntriesApiClient.GetEntriesEnvAsync(request);
+                var request = new GetEntriesEnvRequest(
+                    m_ApiConfig.EnvironmentId.ToString(),
+                    bucket.ToString(),
+                    m_ApiConfig.ProjectId.ToString(),
+                    page,
+                    perPage: entriesPerPage);
+                try
+                {
+                    res = await m_EntriesApiClient.GetEntriesEnvAsync(request);
+                }
+                catch (HttpException exp) when (exp.Response.StatusCode ==
+                                                (int)HttpStatusCode.RequestedRangeNotSatisfiable)
+                {
+                    //page ends exactly at a multiple of 100.
+                    break;
+                }
 
                 foreach (var entry in res.Result)
                 {