+++ Editor/Authoring/CloudContent/ICcdApiConfig.cs
@@ -1,8 +1,0 @@
-namespace Unity.Services.Multiplay.Authoring.Editor.CloudContent
-{
-    interface ICcdApiConfig
-    {
-        string ProjectId { get; }
-        string EnvironmentId { get; }
-    }
-}