+++ Runtime/Authoring/Core/Assets/FleetName.cs
@@ -1,20 +1,33 @@
 namespace Unity.Services.Multiplay.Authoring.Core.Assets
 {
-    readonly struct FleetName : IResourceName
+    /// <summary>
+    /// Struct representing a type-safe Fleet Identifier
+    /// </summary>
+    public readonly struct FleetName : IResourceName
     {
+        /// <summary>
+        /// Creates a new instance of FleetName
+        /// </summary>
         public FleetName(string name) { Name = name; }
+
+        /// <summary>
+        /// The name of the Fleet
+        /// </summary>
         public string Name { get; init; }
 
+        /// <inheritdoc />
         public override string ToString()
         {
             return Name;
         }
 
+        /// <inheritdoc />
         public override bool Equals(object obj)
         {
             return obj is FleetName name && name.Name == Name;
         }
 
+        /// <inheritdoc />
         public override int GetHashCode()
         {
             return Name.GetHashCode();