+++ Runtime/Authoring/Core/Builds/BuildEntry.cs
@@ -1,4 +1,3 @@
-using System;
 using System.IO;
 
 namespace Unity.Services.Multiplay.Authoring.Core.Builds