+++ Runtime/Authoring/Core/MultiplayApi/BuildId.cs
@@ -1,9 +1,13 @@
 namespace Unity.Services.Multiplay.Authoring.Core.MultiplayApi
 {
-    struct BuildId
+    /// <summary>
+    /// Represents a type-safe Build ID
+    /// </summary>
+    public struct BuildId
     {
-        public long Id { private get; init; }
-
-        public long ToLong() => Id;
+        /// <summary>
+        /// The numerical value of the Build
+        /// </summary>
+        public long Id { get; init; }
     }
 }