+++ Runtime/Authoring/Core/MultiplayApi/IServersApi.cs
@@ -1,14 +1,26 @@
-using System;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 
 namespace Unity.Services.Multiplay.Authoring.Core.MultiplayApi
 {
-    interface IServersApi
+    /// <summary>
+    /// Provides access to server information for the project.
+    /// See https://services.docs.unity.com/multiplay-config/v1/#tag/Servers for details
+    /// </summary>
+    public interface IServersApi : IInitializable
     {
+        /// <summary>
+        /// Trigger an action against the server with the given id
+        /// </summary>
         Task<bool> TriggerServerActionAsync(long serverId, ServerAction action, CancellationToken cancellationToken = default);
+        /// <summary>
+        /// Get the details of a single server with the given ID
+        /// </summary>
         Task<ServerInfo> GetServerAsync(long serverId, CancellationToken cancellationToken = default);
+        /// <summary>
+        /// Lists the Action Logs for a server
+        /// </summary>
         Task<List<ActionLog>> GetServerActionLogsAsync(long serverId, CancellationToken cancellationToken = default);
     }
 }