+++ Runtime/Authoring/Core/MultiplayApi/IServersApiFactory.cs
@@ -1,9 +1,0 @@
-using System.Threading.Tasks;
-
-namespace Unity.Services.Multiplay.Authoring.Core.MultiplayApi
-{
-    interface IServersApiFactory
-    {
-        Task<IServersApi> Build();
-    }
-}