+++ Runtime/Authoring/Core/MultiplayApi/LogSearchParams.cs
@@ -2,5 +2,5 @@ using System;
 
 namespace Unity.Services.Multiplay.Authoring.Core.MultiplayApi
 {
-    record LogSearchParams(Guid FleetId, long ServerId, string Query, int Size, DateTime From, DateTime To, string PaginationToken = null);
+    public record LogSearchParams(Guid FleetId, long ServerId, string Query, int Size, DateTime From, DateTime To, string PaginationToken = null);
 }